backend/requirements.txt
Django>=5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
docker>=7.0.0
psutil==5.9.8
gunicorn>=21.2.0 